Slight mind test cleanup
