+++ Content.Server/Mind/MindSystem.cs
@@ -390,14 +390,10 @@ public sealed class MindSystem : EntitySystem
 
         if (entity != null)
         {
-            if (!TryComp(entity.Value, out component))
-            {
-                component = AddComp<MindContainerComponent>(entity.Value);
-            }
-            else if (component.HasMind)
-            {
+            component = EnsureComp<MindContainerComponent>(entity.Value);
+
+            if (component.HasMind)
                 _gameTicker.OnGhostAttempt(component.Mind, false);
-            }
 
             if (TryComp<ActorComponent>(entity.Value, out var actor))
             {
